feat: LoginView, LoginCore 생성
