+++ Projects/App/Sources/RootView.swift
@@ -7,6 +7,7 @@
 
 import ComposableArchitecture
 import DesignSystem
+import KakaoLogin
 import SwiftUI
 
 struct RootView: View {
@@ -41,76 +42,11 @@ struct RootView: View {
   }
 }
 
-struct LoginView: View {
-  public let store: StoreOf<RootCore>
-  
-  public init(store: StoreOf<RootCore>) {
-    self.store = store
-  }
-  
-  var body: some View {
-    VStack {
-      Spacer()
-        .frame(height: 50)
-      Image(uiImage: DesignSystem.Icons.Login.logoUIImage)
-        .resizable()
-        .scaledToFit()
-        .frame(width: 100)
-      Spacer()
-        .frame(height: 8)
-      Text("우리가 픽! 하는\n우리끼리 네컷앨범")
-        .font(DesignSystemFontFamily.Pretendard.regular.font(size: 21))
-        .multilineTextAlignment(.center)
-        .foregroundStyle(DesignSystem.Colors.gray80)
-        .padding()
-      Spacer()
-        .frame(height: 400)
-      Button(action: {
-        store.send(.loginButtonTapped)
-      }, label: {
-        Image(uiImage: DesignSystem.Icons.Login.kakaoUIImage
-          .withPadding(.init(
-            top: 0,
-            left: 16,
-            bottom: 0,
-            right: 16)
-          ))
-        .resizable()
-        .scaledToFit()
-        .frame(width: UIScreen.main.bounds.size.width)
-      })
-    }
-  }
-}
-
-struct ToastView: View {
-  let message: String
-  
-  var body: some View {
-    HStack {
-      Image(uiImage: DesignSystem.Icons.Login.loginIconUIImage)
-        .resizable()
-        .frame(width: 20, height: 20)
-        .padding(.leading)
-      Text(message)
-        .foregroundColor(.white)
-        .padding(.trailing)
-    }
-    .padding()
-    .background(DesignSystem.Colors.gray60)
-    .cornerRadius(40)
-  }
-}
-
 #Preview {
-  ZStack {
-    RootView(
-      store: Store(
-        initialState: RootCore.State(),
-        reducer: { RootCore() }
-      )
+  RootView(
+    store: Store(
+      initialState: RootCore.State(),
+      reducer: { RootCore() }
     )
-    ToastView(message: "로그인에 실패했어요.")
-    Spacer()
-  }
+  )
 }
