+++ Projects/Features/Scene/Project.swift
@@ -16,6 +16,19 @@ let project = Project.make(
         .external(externalDependency: .nuke),
         .external(externalDependency: .lottie)
       ]
+    ),
+    .make(
+      name: "KakaoLogin",
+      product: .staticFramework,
+      bundleId: "com.mashup.gabbangzip.main.kakaoLogin",
+      sources: ["MainScene/Main/**"],
+      dependencies: [
+        .project(target: .coreKit, projectPath: .core),
+        .project(target: .designSystem, projectPath: .designSystem),
+        .external(externalDependency: .composableArchitecture),
+        .external(externalDependency: .nuke),
+        .external(externalDependency: .lottie)
+      ]
     )
   ]
 )
